+++ example/lib/main.dart
@@ -55,7 +55,6 @@ class _MyHomePageState extends State<MyHomePage> {
   void _incrementCounter() async {
     _counter++;
     await _localStorage.put(key: 'count', value: _counter);
-
     setState(() {});
   }
 
@@ -67,6 +66,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   void _init() async {
     _localStorage = await LocalStorage.getInstance();
+
     final count = _localStorage.get<int>(key: 'count');
     if (count != null) {
       _counter = count;